Add readHDOP and getAccuracy
Added a solution to issue #15

Added two functions to get information about the current accuracy of the GPS.
Needed for example to do a Kalman filter of the signal or any other
signal processing.

The calculation of the accuracy is not done in any spectacular way,
just a multiplication of the positional accuracy, but it should at least give
a sense of what the accuracy is.